move uploading to separate file, push command
