+++ src/commands/push.ts
@@ -1,12 +1,14 @@
+import { uploader } from '../utils/uploader';
+import { linker } from '../utils/linker';
 import { wrapAsync } from '../utils/wrap-async';
-import { upload } from '../commands/upload';
-import { link } from '../commands/link';
 
 export function push(vorpal: any) {
   vorpal
     .command('push')
     .description('Upload, link IPNS, and link Ethereum in one go.')
-    .action(function(args: any, callback: () => void) {
-      // vorpal.exec('upload').then(() => vorpal.exec('link'));
-    });
+    .action(wrapAsync( async function(args: any, callback: () => void) {
+      await uploader(vorpal, args, callback);
+      await linker(vorpal, args, callback);
+      callback();
+    }));
 }
